Get rid of ConnectCommand
