+++ ViewModels/ModelMatching/SimConnect.cs
@@ -1,6 +1,7 @@
 #nullable enable
 using System;
 using System.Diagnostics;
+using System.Runtime.InteropServices;
 using Microsoft.FlightSimulator.SimConnect;
 
 namespace vmr_generator.ViewModels.ModelMatching
@@ -28,6 +29,41 @@ namespace vmr_generator.ViewModels.ModelMatching
 					_simConnect = value;
 					OnPropertyChanged(nameof(SimConnect));
 				}
+			}
+		}
+
+		const int WM_USER_SIMCONNECT = 0x0402;
+
+		/// <summary>
+		/// Opens a connection to the simulator.
+		/// </summary>
+		/// <param name="handle">The window handle of the app</param>
+		public void ConnectToSim()
+		{
+			if (!IsSimRunning || IsConnected || WindowHandle == IntPtr.Zero)
+			{
+				return;
+			}
+
+			try
+			{
+				SimConnect = new SimConnect("WMR Generator", WindowHandle, WM_USER_SIMCONNECT, null, 0);
+				SimConnect.OnRecvOpen += SimConnect_OnRecvOpen;
+				SimConnect.OnRecvQuit += SimConnect_OnRecvQuit;
+				SimConnect.OnRecvException += SimConnect_OnRecvException;
+				SimConnect.OnRecvEnumerateInputEvents += SimConnect_OnRecvEnumerateInputEvents;
+			}
+			catch (COMException ex)
+			{
+				// This is just when the sim isn't running, no need to warn the user just log
+				// it to debug and return.
+				if (ex.HResult == -2147467259)
+				{
+					Debug.WriteLine($"Error connecting to the simulator, it probably isn't running: {ex.Message}");
+					return;
+				}
+
+				ErrorMessage = String.Format(_resourceManager.GetString("SimConnectionErrror") ?? "", ex.Message);
 			}
 		}
